move const to models
